+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6N5KY0ZBGY6P4YQE7D30BJ4"
+        "01K6W03A427CEDWA0ZYEVKJ4VD"
     )
     r"""The id of the resource"""